requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
psycopg2-binary>=2.9
celery[redis]>=5.4
gunicorn>=22.0
redis>=5.0
drf-spectacular>=0.27  
black>=24.0
flake8>=7.0